moved artists state to artistsection from main, added table borders
