+++ src/components/Main.tsx
@@ -1,5 +1,4 @@
-import ArtistSearchForm from "./Artists/ArtistSearchSection";
-import ArtistResultTable from "./Artists/ArtistResultTable";
+import ArtistSection from "./Artists/ArtistSection";
 import { useState } from "react";
 import { Artist } from "../Types";
 import SongSection from "./Songs/SongSection";
@@ -9,46 +8,17 @@ export type SearchProps = {
 };
 
 const Main = ({ token }: SearchProps) => {
-  const [artists, setArtists] = useState<null | Array<Artist>>(null);
+  
   const [selectedArtist, setSelectedArtist] = useState<null | Artist>(null);
-
-
-  /* TODO: add pagination so users can select between more than the top 5 artists that show up */
-  async function handleArtistSearch(searchKey: string) {
-    const params = new URLSearchParams({
-      q: searchKey,
-      type: "artist",
-      limit: "5",
-    });
-
-    const response = await fetch(
-      `https://api.spotify.com/v1/search?${params}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    const parsedResponse = await response.json();
-    const responseArtists = parsedResponse.artists.items as Array<Artist>;
-
-    setArtists(responseArtists);
-  }
 
   return (
     <>
       <section>
-        <ArtistSearchForm handleArtistSearch={handleArtistSearch} />
-      </section>
-      <section>
-        {artists ? (
-          <ArtistResultTable
-            artists={artists}
-            setSelectedArtist={setSelectedArtist}
-            selectedArtist={selectedArtist}
-          />
-        ) : null}
+        <ArtistSection 
+        selectedArtist={selectedArtist}
+        setSelectedArtist={setSelectedArtist}
+        token={token}
+        />
       </section>
       <section>
         {selectedArtist ? (
